+++ metadata/metadata_service/api/badge.py
@@ -11,7 +11,7 @@ from flask_restful import Resource, fields, marshal
 
 from metadata_service.entity.badge import Badge
 from metadata_service.exception import NotFoundException
-from metadata_service.proxy import get_proxy_client
+from metadata_service.config import get_proxy_client
 from metadata_service.proxy.base_proxy import BaseProxy
 
 badge_fields = {
