+++ metadata/metadata_service/api/column.py
@@ -13,7 +13,7 @@ from flask_restful import Resource, reqparse
 
 from metadata_service.api.badge import BadgeCommon
 from metadata_service.exception import NotFoundException
-from metadata_service.proxy import get_proxy_client
+from metadata_service.config import get_proxy_client
 
 
 class ColumnLineageAPI(Resource):
